requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
requests
PyJWT
bcrypt
openai
aiofiles
google-genai
numpy
scikit-learn
tensorflow
opencv-python
pillow
gunicorn